AESYNC-v1/backend/requirements.txt
flask 
openai-whisper
werkzeug